consolidate unit tests
